przejscie z int na double

+++ MPD_McNaughton_algorithm/Task.cs
@@ -1,20 +1,16 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace MPD_McNaughton_algorithm
 {
-    public class Task: ICloneable
+    public class Task : ICloneable
     {
         public String Name { get; private set; }
-        public Int32 Duration { get; set; }
+        public Double Duration { get; set; }
         public Color Color;
 
 
-        public Task(String name, Int32 duration, Color color)
+        public Task(String name, Double duration, Color color)
         {
             Name = name;
             Duration = duration;
